Applied atan2 normalization and refactored
This commit completely refactored the PID control. Instead of running all processes in one file, each specific operation is given its unique file in a package. This vastly improves code simplicity and readability. In addition, the normalization is now accomplished  through the two-argument arctangent function. This limits the degrees passed by the gyro from -180 to +180.